Add support for long long type in WebIDL (#106)
After this patch, we can use longlong type in bridge classes between
WebIDL and C++ implementation.

ISSUE=#71

+++ examples/calculator.cc
@@ -34,7 +34,7 @@ int16_t Calculator::Sub(int16_t number1, int16_t number2) {
   return number1 - number2;
 }
 
-double Calculator::Mul(double number1, double number2) {
+int64_t Calculator::Mul(int64_t number1, int64_t number2) {
   return number1 * number2;
 }
